+++ src/clj/rems/api/extra_pages.clj
@@ -1,29 +1,31 @@
 (ns rems.api.extra-pages
   (:require [clojure.java.io :as io]
             [compojure.api.sweet :refer :all]
+            [medley.core :refer [find-first]]
             [rems.api.util :as api-util]
             [rems.config :refer [env]]
-            [rems.common.util :refer [index-by]]
+            [rems.common.roles :as roles]
             [ring.util.http-response :refer :all]
-            [schema.core :as s])
-  (:import (java.io FileNotFoundException)))
+            [schema.core :as s]))
 
 (s/defschema ExtraPageResponse
-  {s/Keyword s/Str})
+  {s/Keyword (s/maybe s/Str)})
 
 (defn- get-extra-page [page-id]
-  (let [allowed-ids (index-by [:id] (filter #(not (:url %)) (:extra-pages env)))]
-    (when (contains? allowed-ids page-id)
-      (let [translations (get-in allowed-ids [page-id :translations])
-            extra-pages-path (:extra-pages-path env)]
-        (assert extra-pages-path ":extra-pages-path undefined in config")
-        (into
-         {}
-         (for [[lang {:keys [filename]}] translations]
-           (let [file (io/file extra-pages-path filename)]
-             (if (.isFile file)
-               [lang (slurp file)]
-               (throw (FileNotFoundException. (str "the file specified in extra-pages does not exist: " file)))))))))))
+  (let [extra-pages (:extra-pages env)]
+    (when-let [page (find-first (comp #{page-id} :id) extra-pages)]
+      (let [roles (:roles page)]
+        (when (or (nil? roles) ; default is unlimited
+                  (apply roles/has-roles? roles))
+          (let [extra-pages-path (:extra-pages-path env)]
+            (assert extra-pages-path ":extra-pages-path undefined in config")
+            (into {}
+                  (for [lang (:languages env)
+                        :let [filename (or (get-in page [:translations lang :filename])
+                                           (:filename page))]]
+                    [lang (when filename
+                            (let [file (io/file extra-pages-path filename)]
+                              (when (.isFile file) (slurp file))))]))))))))
 
 (def extra-pages-api
   (context "/extra-pages" []