+++ src/clj/rems/api/public.clj
@@ -13,9 +13,15 @@
 
 (s/defschema ExtraPage
   {:id s/Str
+   (s/optional-key :filename) s/Str
    (s/optional-key :url) s/Str
-   :translations {s/Keyword {:title s/Str
-                             (s/optional-key :filename) s/Str}}})
+   (s/optional-key :roles) #{s/Keyword}
+   (s/optional-key :show-menu) s/Bool
+   (s/optional-key :show-footer) s/Bool
+   (s/optional-key :heading) s/Bool
+   (s/optional-key :translations) {s/Keyword {(s/optional-key :title) s/Str
+                                              (s/optional-key :filename) s/Str
+                                              (s/optional-key :url) s/Str}}})
 
 (s/defschema GetConfigResponse
   {:authentication s/Keyword