+++ src/clj/rems/css/styles.clj
@@ -576,12 +576,13 @@
                :padding-top "1rem"
                :padding-bottom "1rem"
                :background-color (theme-getx :footer-bgcolor :table-heading-bgcolor :color3)
-               :margin-top (u/em 1)}
-      [:a :a:hover {:color footer-text-color
-                    :font-weight (button-navbar-font-weight)}]
-      [:.dev-reload-button {:position :fixed
-                            :bottom 0
-                            :right (u/px 140)}]])
+               :margin-top (u/em 1)
+               :position :relative}
+      [:a :a:hover :.nav-link {:color footer-text-color
+                               :font-weight (button-navbar-font-weight)}]
+      [:.dev-reload-button {:position :absolute
+                            :bottom (u/rem 1.5)
+                            :right (u/rem 1.5)}]])
 
    [:.jumbotron
     {:background-color "#fff"